CMS-459: Calculate and display effectiveness of cache

+++ modules/cms-core/src/main/java/com/enonic/cms/framework/cache/CacheFacade.java
@@ -84,7 +84,7 @@ public interface CacheFacade
     /**
      * Number of times of clear cache.
      */
-    public long getRemoveAllCount();
+    public int getRemoveAllCount();
 
     /**
      * Return the number of objects in cache.
